git-svn-id: https://scm.ti-edu.ch/reposvn/verify_cprover@120 533bbb3b-ce99-0410-9130-e5b668893301

+++ trunk/cprover/src/goto-programs/write_goto_binary.cpp
@@ -59,22 +59,22 @@ bool write_goto_binary_v2(
     write_long(out, sym.ordering);
 
     unsigned flags=0;    
-    flags = (flags << 1) | sym.is_type; 
-    flags = (flags << 1) | sym.theorem;
-    flags = (flags << 1) | sym.is_macro;
-    flags = (flags << 1) | sym.is_exported;
-    flags = (flags << 1) | sym.is_input;
-    flags = (flags << 1) | sym.is_output;
-    flags = (flags << 1) | sym.is_statevar;
-    flags = (flags << 1) | sym.is_actual;
-    flags = (flags << 1) | sym.free_var;
-    flags = (flags << 1) | sym.binding;
-    flags = (flags << 1) | sym.lvalue;
-    flags = (flags << 1) | sym.static_lifetime;
-    flags = (flags << 1) | sym.thread_local;
-    flags = (flags << 1) | sym.file_local;
-    flags = (flags << 1) | sym.is_extern;
-    flags = (flags << 1) | sym.is_volatile;
+    flags = (flags << 1) | (int)sym.is_type; 
+    flags = (flags << 1) | (int)sym.theorem;
+    flags = (flags << 1) | (int)sym.is_macro;
+    flags = (flags << 1) | (int)sym.is_exported;
+    flags = (flags << 1) | (int)sym.is_input;
+    flags = (flags << 1) | (int)sym.is_output;
+    flags = (flags << 1) | (int)sym.is_statevar;
+    flags = (flags << 1) | (int)sym.is_actual;
+    flags = (flags << 1) | (int)sym.free_var;
+    flags = (flags << 1) | (int)sym.binding;
+    flags = (flags << 1) | (int)sym.lvalue;
+    flags = (flags << 1) | (int)sym.static_lifetime;
+    flags = (flags << 1) | (int)sym.thread_local;
+    flags = (flags << 1) | (int)sym.file_local;
+    flags = (flags << 1) | (int)sym.is_extern;
+    flags = (flags << 1) | (int)sym.is_volatile;
     
     write_long(out, flags);
   }
